Minor refactoring of the pre-allocator
